Made architecture better
